sqrt v int

+++ mymath.go
@@ -2,8 +2,8 @@ package mymath
 
 import "math"
 
-func Sqrt(x float64) float64 {
-	return math.Sqrt(x)
+func Sqrt(x float64) int {
+    	return int(math.Sqrt(x))
 }
 
 func Pow(x, y float64) float64 {
